+++ lib/src/components/tabs.dart
@@ -375,7 +375,6 @@ class ShadTab<T> extends StatefulWidget implements PreferredSizeWidget {
     required this.value,
     required this.child,
     this.content,
-    @Deprecated('Use leading instead') this.icon,
     this.leading,
     this.trailing,
     this.enabled = true,
@@ -441,15 +440,6 @@ class ShadTab<T> extends StatefulWidget implements PreferredSizeWidget {
   /// Displayed when the tab is selected.
   /// {@endtemplate}
   final Widget? content;
-
-  /// {@template ShadTab.icon}
-  /// Deprecated, use [leading] instead.
-  ///
-  /// An icon to display in the tab. Consider using [leading] for semantic
-  /// correctness.
-  /// {@endtemplate}
-  @Deprecated('Use leading instead')
-  final Widget? icon;
 
   /// {@template ShadTab.leading}
   /// A widget to display at the start of the tab, often an [Icon].
@@ -856,7 +846,6 @@ class _ShadTabState<T> extends State<ShadTab<T>> {
             .mergeWith(widget.decoration);
 
         return ShadButton.secondary(
-          icon: widget.icon,
           leading: widget.leading,
           trailing: widget.trailing,
           focusNode: focusNode,