+++ lib/src/theme/components/checkbox.dart
@@ -3,7 +3,6 @@ import 'dart:ui';
 import 'package:flutter/widgets.dart';
 
 import 'package:shadcn_ui/src/theme/components/decorator.dart';
-import 'package:shadcn_ui/src/utils/extensions/order_policy.dart';
 
 @immutable
 class ShadCheckboxTheme {
@@ -15,7 +14,6 @@ class ShadCheckboxTheme {
     this.size,
     this.padding,
     this.crossAxisAlignment,
-    this.orderPolicy,
   });
 
   final bool merge;
@@ -38,9 +36,6 @@ class ShadCheckboxTheme {
   /// {@macro ShadCheckbox.crossAxisAlignment}
   final CrossAxisAlignment? crossAxisAlignment;
 
-  /// {@macro ShadCheckbox.orderPolicy}
-  final WidgetOrderPolicy? orderPolicy;
-
   static ShadCheckboxTheme lerp(
     ShadCheckboxTheme a,
     ShadCheckboxTheme b,
@@ -55,7 +50,6 @@ class ShadCheckboxTheme {
       size: lerpDouble(a.size, b.size, t),
       padding: EdgeInsets.lerp(a.padding, b.padding, t),
       crossAxisAlignment: t < .5 ? a.crossAxisAlignment : b.crossAxisAlignment,
-      orderPolicy: t < .5 ? a.orderPolicy : b.orderPolicy,
     );
   }
 
@@ -67,7 +61,6 @@ class ShadCheckboxTheme {
     ShadDecoration? decoration,
     EdgeInsets? padding,
     CrossAxisAlignment? crossAxisAlignment,
-    WidgetOrderPolicy? orderPolicy,
   }) {
     return ShadCheckboxTheme(
       merge: merge ?? this.merge,
@@ -77,7 +70,6 @@ class ShadCheckboxTheme {
       color: color ?? this.color,
       padding: padding ?? this.padding,
       crossAxisAlignment: crossAxisAlignment ?? this.crossAxisAlignment,
-      orderPolicy: orderPolicy ?? this.orderPolicy,
     );
   }
 
@@ -91,7 +83,6 @@ class ShadCheckboxTheme {
       size: other.size,
       padding: other.padding,
       crossAxisAlignment: other.crossAxisAlignment,
-      orderPolicy: other.orderPolicy,
     );
   }
 
@@ -106,8 +97,7 @@ class ShadCheckboxTheme {
         other.duration == duration &&
         other.decoration == decoration &&
         other.padding == padding &&
-        other.crossAxisAlignment == crossAxisAlignment &&
-        other.orderPolicy == orderPolicy;
+        other.crossAxisAlignment == crossAxisAlignment;
   }
 
   @override
@@ -118,7 +108,6 @@ class ShadCheckboxTheme {
         duration.hashCode ^
         decoration.hashCode ^
         padding.hashCode ^
-        crossAxisAlignment.hashCode ^
-        orderPolicy.hashCode;
+        crossAxisAlignment.hashCode;
   }
 }